Adapt dispatch retry logic
